Move test utils in subfolder and split it by subject
