+++ src/test_utils/print_board.py
@@ -1,0 +1,7 @@
+import sys
+from render import print_board
+
+
+def print_board_if_verbosity_is_set(board):
+    if "-v" in sys.argv:
+        print_board(board, False)
